XML/Python Code for the project
The eye detection.py is the one that is run on a computer that has a webcam attached to it. that is the program that will take images and process the eyes direction and blinks. You need to have the xml files in the same location as the .py file for it to work correctly.

The socket code is ran on the raspberry pi. that is the code that accepts the socket command and in turn runs the associated motors for the car.